+++ src/main/java/com/dabomstew/pkrandom/romhandlers/Gen5RomHandler.java
@@ -1168,6 +1168,7 @@ public class Gen5RomHandler extends AbstractDSRomHandler {
             getTextHandler().bw1CasteliaCityBurghTextModifications(taggedTypes);
             getTextHandler().bw1CherenBurghTextModifications(taggedTypes);
             getTextHandler().bw1NimbasaCityTextModifications(taggedTypes, this.random);
+            getTextHandler().bw1JuniperTextModifications(taggedTypes);
         }
     }