+++ src/test/java/com/dabomstew/pkrandom/romhandlers/Gen5Test.java
@@ -213,7 +213,31 @@ public class Gen5Test {
         // Verify with type theming
         romhandler.typeThemeTrainerPokes(false, false, false, false, false, 0);
         verify(mockTextHandler, times(1)).bw1NimbasaCityTextModifications(isNotNull(), any());
-    }    
+    }  
+    
+    /**
+     * When altering trainers, make sure the ROM's text is updated when it is a
+     * Black/White ROM
+     * 
+     * @throws IOException
+     */
+    @Test
+    public void TestBW1JuniperElesaText() throws IOException {
+        // Test with and without null
+        Gen5RomHandler romhandler = spy(new Gen5RomHandler(new Random()));
+        resetDataModel(romhandler);
+        doReturn(Gen5RomHandler.getRomFromSupportedRom("Black (U)")).when(romhandler).getRomEntry();
+        doReturn(0xFFFF).when(romhandler).readWord(any(), anyInt());
+        doReturn(pokemonList.subList(0, 3)).when(romhandler).getStarters();
+        
+        // Verify with random team
+        romhandler.randomizeTrainerPokes(false, false, false, 0);
+        verify(mockTextHandler, times(1)).bw1JuniperTextModifications(isNull());
+
+        // Verify with type theming
+        romhandler.typeThemeTrainerPokes(false, false, false, false, false, 0);
+        verify(mockTextHandler, times(1)).bw1JuniperTextModifications(isNotNull());
+    }
 
     /**
      * Function for granular modification of data model